stripe_payment/payment/api/v1/serializers.py
from rest_framework import serializers
from payment.models import User

class LoginSerializer(serializers.ModelSerializer):
    username = serializers.CharField()
    password = serializers.CharField()

    class Meta:
        model = User
        fields = ['username', 'password']


class CardDetailSerializer(serializers.Serializer):
    card_number = serializers.CharField(required=True)
    card_holder = serializers.CharField(required=True)
    card_exp_month = serializers.CharField(required=True)
    card_exp_year = serializers.CharField(required=True)
    card_cvv = serializers.CharField(required=True)
    amount = serializers.IntegerField(help_text="Provide amount in cents", required=True)


class BankDetailSerializer(serializers.Serializer):
    country = serializers.CharField(required=True)
    currency = serializers.CharField(required=True)
    account_holder_name = serializers.CharField(required=True)
    account_holder_type = serializers.CharField(required=True)
    routing_number = serializers.CharField(required=True)
    account_number = serializers.CharField(required=True)
    amount = serializers.IntegerField(help_text="Provide amount in cents", required=True)



