stripe_payment/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
dj-stripe==2.5.1
Django==3.2.8
djangorestframework==3.12.4
idna==3.3
jsonfield==3.1.0
Pillow==8.4.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
stripe==2.61.0
typing-extensions==3.10.0.2
urllib3==1.26.7
